clean up for build to run
